Mejora de clasificación

+++ data/list_modebase.py
@@ -2,4 +2,4 @@ from pydantic import BaseModel, BaseSettings
 from typing import List
 
 class values_list(BaseModel):
-    data_model : list[float]
+    data_model : list[int]
